Refactor frontend controller organization
