+++ erun-ui/frontend/src/app/model/appStatusPayload.ts
@@ -1,0 +1,4 @@
+export interface AppStatusPayload {
+  message?: string;
+  busy?: boolean;
+}
