+++ erun-ui/frontend/src/app/model/debugOpenFilter.ts
@@ -1,0 +1,4 @@
+export interface DebugOpenFilter {
+  released: boolean;
+  pending: string;
+}
