+++ erun-ui/frontend/src/app/model/debugSessionMode.ts
@@ -1,0 +1,1 @@
+export type DebugSessionMode = 'open' | 'hidden';
