+++ erun-ui/frontend/src/app/model/hiddenSessionMode.ts
@@ -1,0 +1,1 @@
+export type HiddenSessionMode = 'sshd-init' | 'doctor';
