+++ erun-ui/frontend/src/app/model/ideKind.ts
@@ -1,0 +1,1 @@
+export type IDEKind = 'vscode' | 'intellij';
