+++ erun-ui/frontend/src/app/model/idleCloudContextAction.ts
@@ -1,0 +1,9 @@
+import type { UIIdleStatus } from '@/types';
+
+export type IdleCloudContextAction = {
+  idleStatus: UIIdleStatus;
+  name: string;
+  run: (name: string) => Promise<unknown>;
+  label: string;
+  refreshKubernetesContexts: boolean;
+};
