+++ erun-ui/frontend/src/app/model/mountElements.ts
@@ -1,0 +1,7 @@
+export interface MountElements {
+  terminalRoot: HTMLDivElement;
+  terminalPane: HTMLElement;
+  reviewView: HTMLElement;
+  reviewMain: HTMLDivElement;
+  diffList: HTMLDivElement;
+}
