+++ erun-ui/frontend/src/app/model/normalizedEnvironmentDialogValues.ts
@@ -1,0 +1,7 @@
+export type NormalizedEnvironmentDialogValues = {
+  tenant: string;
+  environment: string;
+  version: string;
+  kubernetesContext: string;
+  containerRegistry: string;
+};
