+++ erun-ui/frontend/src/app/model/terminalDataDisposable.ts
@@ -1,0 +1,3 @@
+import type { Terminal } from '@xterm/xterm';
+
+export type TerminalDataDisposable = ReturnType<Terminal['onData']>;
