+++ erun-ui/frontend/src/app/model/terminalExitSelections.ts
@@ -1,0 +1,10 @@
+import type { UISelection } from '@/types';
+
+export type TerminalExitSelections = {
+  initSelection?: UISelection;
+  deploySelection?: UISelection;
+  sshdInitSelection?: UISelection;
+  doctorSelection?: UISelection;
+  openSelection?: UISelection;
+  cloudInit: boolean;
+};
