+++ erun-ui/frontend/src/app/model/terminalWriteData.ts
@@ -1,0 +1,1 @@
+export type TerminalWriteData = string | Uint8Array;
