+++ erun-ui/frontend/src/app/sessionSelection.ts
@@ -1,0 +1,3 @@
+export function isNewSessionSelection(previousSessionId: number, previousKnownSessionId: number): boolean {
+  return previousKnownSessionId === 0 || previousKnownSessionId !== previousSessionId;
+}
